+++ examples/titanic_example/assets/dataset/titanic_opener.py
@@ -16,13 +16,6 @@ class TitanicOpener(tools.Opener):
         data = self._get_data(folders)
         return self._get_y(data)
 
-    def save_predictions(self, y_pred, path):
-        with open(path, "w") as f:
-            y_pred.to_csv(f, index=False)
-
-    def get_predictions(self, path):
-        return pd.read_csv(path)
-
     def fake_X(self, n_samples=None):
         data = self._fake_data(n_samples)
         return self._get_X(data)
@@ -47,23 +40,14 @@ class TitanicOpener(tools.Opener):
             "PassengerId": list(range(N_SAMPLES)),
             "Survived": [random.choice([True, False]) for k in range(N_SAMPLES)],
             "Pclass": [random.choice([1, 2, 3]) for k in range(N_SAMPLES)],
-            "Name": [
-                "".join(random.sample(string.ascii_letters, 10))
-                for k in range(N_SAMPLES)
-            ],
+            "Name": ["".join(random.sample(string.ascii_letters, 10)) for k in range(N_SAMPLES)],
             "Sex": [random.choice(["male", "female"]) for k in range(N_SAMPLES)],
             "Age": [random.choice(range(7, 77)) for k in range(N_SAMPLES)],
             "SibSp": [random.choice(range(4)) for k in range(N_SAMPLES)],
             "Parch": [random.choice(range(4)) for k in range(N_SAMPLES)],
-            "Ticket": [
-                "".join(random.sample(string.ascii_letters, 10))
-                for k in range(N_SAMPLES)
-            ],
+            "Ticket": ["".join(random.sample(string.ascii_letters, 10)) for k in range(N_SAMPLES)],
             "Fare": [random.choice(np.arange(15, 150, 0.01)) for k in range(N_SAMPLES)],
-            "Cabin": [
-                "".join(random.sample(string.ascii_letters, 3))
-                for k in range(N_SAMPLES)
-            ],
+            "Cabin": ["".join(random.sample(string.ascii_letters, 3)) for k in range(N_SAMPLES)],
             "Embarked": [random.choice(["C", "S", "Q"]) for k in range(N_SAMPLES)],
         }
         return pd.DataFrame(data)
@@ -73,11 +57,7 @@ class TitanicOpener(tools.Opener):
         # find csv files
         paths = []
         for folder in folders:
-            paths += [
-                os.path.join(folder, f)
-                for f in os.listdir(folder)
-                if f.endswith(".csv")
-            ]
+            paths += [os.path.join(folder, f) for f in os.listdir(folder) if f.endswith(".csv")]
 
         # load data
         data = pd.DataFrame()