+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* chore: change the assets and tools algo to feed with inputs outputs dictionary. (#113)
+
 ## [0.20.0]
 
 * doc: add GPU page on substra documentation